Opening the TZ database once. This feels a bit risky?
It essentially will tie up a bunch of memory in the main thread. Maybe
we're better off opening/closing the database every time? Easy to
go from one pattern to the other, I _think_ it's safe because the
ZoneDetect library doesn't seem to have any thread-unsafe code going on,
but we'll see!