Sacrifice foresight for precision
The immediate crossings notion was weird. We now simply
see if there's crossings in the bracket of 24 hours right
before or after the moment of query; vs. the latest of
the found crossing candidates -- this was proving to be
wildly inaccurate in the general case!

Still cool to see when crossings may happen though.